+++ gnat/connectors/group_ib/__init__.py
@@ -1,0 +1,3 @@
+from .client import GroupIBClient
+
+__all__ = ["GroupIBClient"]