+++ gnat/connectors/orca/__init__.py
@@ -1,0 +1,3 @@
+from .client import OrcaClient
+
+__all__ = ["OrcaClient"]